Route requests with query parameters correctly
Currently, when a path based route is established and then requested with
query parameters, a 404 is returned. For example, a registration for
test/my_path with a GET /my_path?foo=bar as the request will result in a
404.

This change modifies the proxy to strip all parts of the query string from
the RequestURI.

Go 1.5 may offer a more succinct fix via request.URL.EscapedPath().